+++ netbox_branching/constants.py
@@ -9,3 +9,9 @@ BRANCH_HEADER = 'X-NetBox-Branch'
 
 # URL query parameter name
 QUERY_PARAM = '_branch'
+
+# Apps which are explicitly excluded from branching
+EXCLUDED_APPS = (
+    'netbox_branching',
+    'netbox_changes',
+)