+++ netbox_branching/migrations/0001_initial.py
@@ -13,7 +13,7 @@ class Migration(migrations.Migration):
     dependencies = [
         ('contenttypes', '0002_remove_content_type_name'),
         ('core', '0011_move_objectchange'),
-        ('extras', '0118_notifications'),
+        ('extras', '0119_notifications'),
         migrations.swappable_dependency(settings.AUTH_USER_MODEL),
     ]