+++ netbox_branching/signal_receivers.py
@@ -1,3 +1,4 @@
+import logging
 from functools import partial
 
 from django.db import DEFAULT_DB_ALIAS
@@ -10,6 +11,7 @@ from core.choices import ObjectChangeActionChoices
 from core.models import ObjectChange, ObjectType
 from extras.events import process_event_rules
 from extras.models import EventRule
+from netbox.registry import registry
 from utilities.exceptions import AbortRequest
 from utilities.serialization import serialize_object
 from .choices import BranchStatusChoices
@@ -30,9 +32,15 @@ def record_change_diff(instance, **kwargs):
     """
     When an ObjectChange is created, create or update the relevant ChangeDiff for the active Branch.
     """
+    logger = logging.getLogger('netbox_branching.signal_receivers.record_change_diff')
+
     branch = active_branch.get()
     object_type = instance.changed_object_type
     object_id = instance.changed_object_id
+
+    # If this type of object does not support branching, return immediately.
+    if object_type.model not in registry['model_features']['branching'].get(object_type.app_label, []):
+        return
 
     # If this is a global change, update the "current" state in any ChangeDiffs for this object.
     if branch is None:
@@ -41,27 +49,31 @@ def record_change_diff(instance, **kwargs):
         if instance.action == ObjectChangeActionChoices.ACTION_CREATE:
             return
 
-        print(f"Updating change diff for global change to {instance.changed_object}")
-        if diff := ChangeDiff.objects.filter(object_type=object_type, object_id=object_id, branch__status=BranchStatusChoices.READY).first():
-            diff.last_updated = timezone.now()
-            diff.current = instance.postchange_data_clean
-            diff.save()
+        logger.debug(f"Updating change diff for global change to {instance.changed_object}")
+        ChangeDiff.objects.filter(
+            object_type=object_type,
+            object_id=object_id,
+            branch__status=BranchStatusChoices.READY
+        ).update(
+            last_updated=timezone.now(),
+            current=instance.postchange_data_clean or None
+        )
 
     # If this is a branch-aware change, create or update ChangeDiff for this object.
     else:
 
         # Updating the existing ChangeDiff
         if diff := ChangeDiff.objects.filter(object_type=object_type, object_id=object_id, branch=branch).first():
-            print(f"Updating branch change diff for change to {instance.changed_object}")
+            logger.debug(f"Updating branch change diff for change to {instance.changed_object}")
             diff.last_updated = timezone.now()
             if diff.action != ObjectChangeActionChoices.ACTION_CREATE:
                 diff.action = instance.action
-            diff.modified = instance.postchange_data_clean
+            diff.modified = instance.postchange_data_clean or None
             diff.save()
 
         # Creating a new ChangeDiff
         else:
-            print(f"Creating branch change diff for change to {instance.changed_object}")
+            logger.debug(f"Creating branch change diff for change to {instance.changed_object}")
             if instance.action == ObjectChangeActionChoices.ACTION_CREATE:
                 current_data = None
             else:
@@ -70,12 +82,11 @@ def record_change_diff(instance, **kwargs):
                 current_data = serialize_object(obj, exclude=['created', 'last_updated'])
             diff = ChangeDiff(
                 branch=branch,
-                object_type=instance.changed_object_type,
-                object_id=instance.changed_object_id,
+                object=instance.changed_object,
                 action=instance.action,
-                original=instance.prechange_data_clean,
-                modified=instance.postchange_data_clean,
-                current=current_data,
+                original=instance.prechange_data_clean or None,
+                modified=instance.postchange_data_clean or None,
+                current=current_data or None,
                 last_updated=timezone.now(),
             )
             diff.save()
@@ -85,6 +96,9 @@ def handle_branch_event(event_type, branch, user=None, **kwargs):
     """
     Process any EventRules associated with branch events (e.g. syncing or merging).
     """
+    logger = logging.getLogger('netbox_branching.signal_receivers.handle_branch_event')
+    logger.debug(f"Checking for {event_type} event rules")
+
     # Find any EventRules for this event type
     object_type = ObjectType.objects.get_by_natural_key('netbox_branching', 'branch')
     event_rules = EventRule.objects.filter(
@@ -92,6 +106,10 @@ def handle_branch_event(event_type, branch, user=None, **kwargs):
         enabled=True,
         object_types=object_type
     )
+    if not event_rules:
+        logger.debug("No matching event rules found")
+        return
+    logger.debug(f"Found {len(event_rules)} event rules")
 
     # Serialize the branch & process EventRules
     username = user.username if user else None